Time: 68 ms (29.99%) | Memory: 47.6 MB (29.04%) - LeetSync

+++ 560-subarray-sum-equals-k/subarray-sum-equals-k.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int subarraySum(vector<int>& nums, int k) {
-        long long n=nums.size(), cnt=0, sum=0;
+        int n=nums.size(), cnt=0, sum=0;
         map<int,int> mp;
         mp[0]=1;
         for(int i=0;i<n;i++){
